requirements.txt
fastapi==0.116.1
joblib==1.5.1
numpy==2.3.2
pandas==2.3.1
pydantic==2.11.7
python-dotenv==1.1.1
Requests==2.32.4
scikit_learn==1.7.1
streamlit==1.48.0
supabase==2.18.0
uvicorn==0.35.0
